worked on Add to cart - sides, dips, drinks menu & Order Place

+++ src/Pages/ongoingOrders/OngoingOrder.js
@@ -7,6 +7,7 @@ import {
   allIngredientsApi,
   deleteCartItemApi,
   getCartListApi,
+  orderPlaceApi,
   sidesApi,
   storeLocationApi,
 } from "../../API/ongoingOrder";
@@ -14,6 +15,9 @@ import SidesMenu from "./SidesMenu";
 import DipsMenu from "./DipsMenu";
 import DrinksMenu from "./DrinksMenu";
 import $ from "jquery";
+import swal from "sweetalert";
+import { ToastContainer, toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 
 function OngoingOrder() {
   const [allIngredients, setAllIngredients] = useState();
@@ -38,6 +42,7 @@ function OngoingOrder() {
         console.log("Error From All Ingredient API: ", err);
       });
   };
+
   //API - Sides Ingredients
   const sidesIngredient = async () => {
     await sidesApi()
@@ -48,6 +53,7 @@ function OngoingOrder() {
         console.log("Error From All Ingredient API: ", err);
       });
   };
+
   //API - Store Location
   const storeLocation = async () => {
     await storeLocationApi()
@@ -86,23 +92,44 @@ function OngoingOrder() {
     }
   };
 
+  //API - Delete Item
   const deleteCartItem = async (e, cartLineCode) => {
     e.preventDefault();
-    let cart = JSON.parse(localStorage.getItem("CartData"));
+    swal({
+      title: "Are you sure you want to delete?",
+      text: "You will not be able to recover this item!",
+      icon: "warning",
+      buttons: ["Cancel", "Delete"],
+      dangerMode: true,
+    }).then(async (willDelete) => {
+      if (willDelete) {
+        let cart = JSON.parse(localStorage.getItem("CartData"));
+        let payload = {
+          cartLineCode: cartLineCode,
+          discountAmount: "0",
+          taxPer: "0",
+        };
+        console.log("Delelte Item", payload);
+        await deleteCartItemApi(payload)
+          .then((res) => {
+            toast.success("Product Deleted Successfully..");
+            getCartList();
+          })
+          .catch((err) => {
+            console.log("Error From Delete Cart Item API: ", err);
+          });
+      }
+    });
+  };
 
-    let payload = {
-      cartLineCode: cartLineCode,
-      discountAmount: "0",
-      taxPer: "0",
-    };
-    console.log("Delelte Item", payload);
-    await deleteCartItemApi(payload)
+  const handleOrderPlace = async (e, cartCode) => {
+    e.preventDefault();
+    await orderPlaceApi({ cartCode: cartCode })
       .then((res) => {
-        console.log(res);
-        getCartList();
+        toast.success(res.data.message);
       })
       .catch((err) => {
-        console.log("Error From Delete Cart Item API: ", err);
+        console.log("ERROR From Order Place API : ", err);
       });
   };
 
@@ -116,11 +143,10 @@ function OngoingOrder() {
   return (
     <>
       <Nav />
-
       <div className="my-2 mx-4">
         <div className="row gx-4">
           {/* Section 1 */}
-          <div className="col-lg-3">
+          <div className="col-lg-2">
             <form>
               <label className="form-label mt-2 mb-1">Phone</label>
               <input
@@ -128,14 +154,8 @@ function OngoingOrder() {
                 type="tel"
                 onChange={(e) => setMobileNumber(e.target.value)}
               />
-              <label className="form-label my-1">Customer Name</label>
-              <input
-                className="form-control"
-                type="text"
-                onChange={(e) => setCustomerName(e.target.value)}
-              />
-              <div className="m-0 p-0">
-                <label className="radio-inline mt-3">
+              <div className="my-3 m-0 p-0 d-flex justify-content-between">
+                <label className="radio d-flex align-items-center my-1 w-50">
                   <input
                     className="mx-2"
                     type="radio"
@@ -147,11 +167,17 @@ function OngoingOrder() {
                   />
                   Pickup
                 </label>
-                <label className="radio-inline mx-4">
+                <label className="radio d-flex align-items-center my-1 mx-2 w-50">
                   <input className="mx-2" type="radio" name="category" />
                   Delivery
                 </label>
               </div>
+              <label className="form-label">Customer Name</label>
+              <input
+                className="form-control"
+                type="text"
+                onChange={(e) => setCustomerName(e.target.value)}
+              />
               <label className="form-label mt-2 mb-1">Address</label>
               <textarea
                 className="form-control"
@@ -182,9 +208,6 @@ function OngoingOrder() {
               <table className="table text-center border-none">
                 <thead>
                   <tr>
-                    <th scope="col">
-                      <input type="checkbox" />
-                    </th>
                     <th scope="col">Date</th>
                     <th scope="col">Order</th>
                   </tr>
@@ -290,7 +313,16 @@ function OngoingOrder() {
 
               {/* All SIdes */}
               <div id="sidesMenu" className="container tab-pane m-0 p-0">
-                <SidesMenu />
+                <SidesMenu
+                  getCartList={getCartList}
+                  customerName={customerName}
+                  mobileNumber={mobileNumber}
+                  address={address}
+                  deliveryType={deliveryType}
+                  storeLocation={storesCode}
+                  discount={discount}
+                  taxPer={taxPer}
+                />
               </div>
 
               {/* All Dips */}
@@ -298,18 +330,36 @@ function OngoingOrder() {
                 id="dipsMenu"
                 className="container tab-pane m-0 p-0 topping-list"
               >
-                <DipsMenu />
+                <DipsMenu
+                  getCartList={getCartList}
+                  customerName={customerName}
+                  mobileNumber={mobileNumber}
+                  address={address}
+                  deliveryType={deliveryType}
+                  storeLocation={storesCode}
+                  discount={discount}
+                  taxPer={taxPer}
+                />
               </div>
 
               {/* All Drinks */}
               <div id="drinksMenu" className="container tab-pane m-0 p-0">
-                <DrinksMenu />
+                <DrinksMenu
+                  getCartList={getCartList}
+                  customerName={customerName}
+                  mobileNumber={mobileNumber}
+                  address={address}
+                  deliveryType={deliveryType}
+                  storeLocation={storesCode}
+                  discount={discount}
+                  taxPer={taxPer}
+                />
               </div>
             </div>
           </div>
 
           {/* Section 3 */}
-          <div className="col-lg-3 my-1">
+          <div className="col-lg-4 my-1">
             <h6 className="text-center">Order</h6>
             {/* Add to Cart */}
             <div className="p-3 rounded mb-3 addToCartList">
@@ -319,11 +369,21 @@ function OngoingOrder() {
                   <div>
                     <div className="d-flex justify-content-between">
                       <h6>{data.productName}</h6>
-                      <span className="mx-4">${data.price}</span>
+                      <span className="mx-0">${data.amount}</span>
                     </div>
-                    <div className="d-flex justify-content-left mb-1">
-                      <h6>Size : </h6>
-                      <span className="mx-1">{data.pizzaSize}</span>
+                    <div className="d-flex justify-content-between">
+                      <div className="d-flex justify-content-left">
+                        <h6>Size : </h6>
+                        <span className="mx-1">
+                          {data.productType === "side"
+                            ? data.config.sidesSize
+                            : data.pizzaSize}
+                        </span>
+                      </div>
+                      <div className="d-flex justify-content-right mx-0 mb-1">
+                        <h6 className="mx-2">Quantity : </h6>
+                        <span className="">{data.quantity}</span>
+                      </div>
                     </div>
                     <div className="d-flex align-items-center">
                       <button
@@ -352,14 +412,38 @@ function OngoingOrder() {
             {/* Order Submit */}
             <div className="my-1">
               <form>
+                {/* Price / Sub Total */}
                 <div className="d-flex flex-wrap my-2 justify-content-end align-items-center OrderAmount">
                   <label className="form-label w-25">Price</label>
-                  <div className="input-group w-50">
+                  <div className="input-group w-75">
                     <div class="input-group-prepend">
                       <span className="input-group-text inputGroupTxt">$</span>
                     </div>
                     <input
-                      className="form-control w-25"
+                      className="form-control w-25 text-end"
+                      type="number"
+                      placeholder="0.00"
+                      step="0.01"
+                      readOnly
+                      value={cartListData?.subTotal}
+                    ></input>
+                    <div class="input-group-append">
+                      <span className="input-group-text inputGroupTxt">
+                        CAD
+                      </span>
+                    </div>
+                  </div>
+                </div>
+
+                {/* Discount */}
+                <div className="d-flex flex-wrap my-2 my-2 justify-content-end align-items-center">
+                  <label className="form-label w-25">Discount</label>
+                  <div className="input-group w-75">
+                    <div class="input-group-prepend">
+                      <span className="input-group-text inputGroupTxt">$</span>
+                    </div>
+                    <input
+                      className="form-control w-25 text-end"
                       type="number"
                       placeholder="0.00"
                       step="0.01"
@@ -367,20 +451,21 @@ function OngoingOrder() {
                     ></input>
                     <div class="input-group-append">
                       <span className="input-group-text inputGroupTxt">
-                        USD
+                        CAD
                       </span>
                     </div>
                   </div>
                 </div>
 
-                <div className="d-flex flex-wrap my-2 my-2 justify-content-end align-items-center">
-                  <label className="form-label w-25">Discount</label>
-                  <div className="input-group w-50">
+                {/* Tax Percentage */}
+                <div className="d-flex flex-wrap my-2 justify-content-end align-items-center">
+                  <label className="form-label w-25">Tax</label>
+                  <div className="input-group w-75">
                     <div class="input-group-prepend">
                       <span className="input-group-text inputGroupTxt">$</span>
                     </div>
                     <input
-                      className="form-control w-25"
+                      className="form-control w-25 text-end"
                       type="number"
                       placeholder="0.00"
                       step="0.01"
@@ -388,56 +473,41 @@ function OngoingOrder() {
                     ></input>
                     <div class="input-group-append">
                       <span className="input-group-text inputGroupTxt">
-                        USD
+                        CAD
                       </span>
                     </div>
                   </div>
                 </div>
 
-                <div className="d-flex flex-wrap my-2 justify-content-end align-items-center">
-                  <label className="form-label w-25">Tax</label>
-                  <div className="input-group w-50">
-                    <div class="input-group-prepend">
-                      <span className="input-group-text inputGroupTxt">$</span>
-                    </div>
-                    <input
-                      className="form-control w-25"
-                      type="number"
-                      placeholder="0.00"
-                      step="0.01"
-                    ></input>
-                    <div class="input-group-append">
-                      <span className="input-group-text inputGroupTxt">
-                        USD
-                      </span>
-                    </div>
-                  </div>
-                </div>
-
+                {/* Grand Total / Total Price */}
                 <div className="d-flex flex-wrap my-2 justify-content-end align-items-center">
                   <label className="form-label w-25">Total Price</label>
-                  <div className="input-group w-50">
+                  <div className="input-group w-75">
                     <div class="input-group-prepend">
                       <span className="input-group-text inputGroupTxt">$</span>
                     </div>
                     <input
-                      className="form-control w-25"
+                      className="form-control w-25 text-end"
                       type="number"
                       placeholder="0.00"
                       step="0.01"
+                      value={cartListData?.grandTotal}
+                      readOnly
                     ></input>
                     <div class="input-group-append">
                       <span className="input-group-text inputGroupTxt">
-                        USD
+                        CAD
                       </span>
                     </div>
                   </div>
                 </div>
 
+                {/* Submit Order */}
                 <div className="d-flex flex-row justify-content-end align-items-center">
                   <button
                     type="submit"
                     className="submitOrderbtn btn btn-sm my-3 px-4 py-2"
+                    onClick={(e) => handleOrderPlace(e, cartListData.code)}
                   >
                     Submit Order
                   </button>
@@ -447,6 +517,7 @@ function OngoingOrder() {
           </div>
         </div>
       </div>
+      <ToastContainer position="top-center" />
     </>
   );
 }
